+++ crates/vm/levm/src/db/gen_db.rs
@@ -386,12 +386,7 @@ impl<'a> VM<'a> {
         key: H256,
     ) -> Result<(U256, bool), InternalError> {
         // [EIP-2929] - Introduced conditional tracking of accessed storage slots for Berlin and later specs.
-        let storage_slot_was_cold = self
-            .substate
-            .accessed_storage_slots
-            .entry(address)
-            .or_default()
-            .insert(key);
+        let storage_slot_was_cold = !self.substate.add_accessed_slot(address, key);
 
         let storage_slot = self.get_storage_value(address, key)?;