+++ crates/vm/levm/src/hooks/default_hook.rs
@@ -222,12 +222,16 @@ pub fn pay_coinbase(vm: &mut VM<'_>, gas_to_pay: u64) -> Result<(), VMError> {
 
 // In Cancun the only addresses destroyed are contracts created in this transaction
 pub fn delete_self_destruct_accounts(vm: &mut VM<'_>) -> Result<(), VMError> {
-    let selfdestruct_set = vm.substate.selfdestruct_set.clone();
-    for address in selfdestruct_set {
-        let account_to_remove = vm.get_account_mut(address)?;
+    for address in vm.substate.iter_selfdestruct() {
+        let account_to_remove = vm.db.get_account_mut(*address)?;
+        vm.current_call_frame
+            .call_frame_backup
+            .backup_account_info(*address, account_to_remove)?;
+
         *account_to_remove = LevmAccount::default();
-        vm.db.destroyed_accounts.insert(address);
+        vm.db.destroyed_accounts.insert(*address);
     }
+
     Ok(())
 }